refactoring Signature and PublicKey in data subpackage
