+++ plugin-api/src/main/java/org/hyperledger/besu/plugin/services/securitymodule/SecurityModule.java
@@ -14,6 +14,9 @@
  */
 package org.hyperledger.besu.plugin.services.securitymodule;
 
+import org.hyperledger.besu.plugin.services.securitymodule.data.PublicKey;
+import org.hyperledger.besu.plugin.services.securitymodule.data.Signature;
+
 import org.apache.tuweni.bytes.Bytes32;
 
 /**
@@ -25,7 +28,7 @@ public interface SecurityModule {
 
   /**
    * @param dataHash The Keccack hash of a set of data, which is to be signed.
-   * @return the signature (R, S, recId) generated by signing the hash with the node key
+   * @return the signature (R, S) generated by signing the hash with the node key
    */
   Signature sign(Bytes32 dataHash);
