+++ plugin-api/src/main/java/org/hyperledger/besu/plugin/services/securitymodule/data/PublicKey.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.plugin.services.securitymodule;
+package org.hyperledger.besu.plugin.services.securitymodule.data;
 
 import org.apache.tuweni.bytes.Bytes;
